+++ CHANGES.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* New music music by [Stey](https://soundcloud.com/stey_music)
 * Add 50 new playable maps
 * Scroll one page at a time in the level select screen with `page up` and `page down`
 * Draw island id and ARtB on the island previews when in map editor mode
@@ -55,6 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Harden against exceptions within a start/end block
 
 ### Removed
+
+* All old music tracks have been removed
 
 ---